Adicionados os Services, e o nosso controller, alem das dependencias @Valid e OpenAPi 3

+++ src/main/java/com/logistica/entregas_api/dto/EntregaRequestDTO.java
@@ -13,7 +13,7 @@ public class EntregaRequestDTO {
 
     public EntregaRequestDTO() {}
 
-    public Integer getClienteId() { return clienteId; }
+    public Long getClienteId() { return clienteId; }
     public void setClienteId(Integer clienteId) { this.clienteId = clienteId; }
 
     public String getCep() { return cep; }
